scrape_menus/requirements.txt
playwright
llama-cpp-python
rich
numpy
openai
python-dotenv
requests
Pillow
transformers

